requirements.txt
-i https://pypi.org/simple
anyio==4.6.2.post1; python_version >= '3.9'
certifi==2024.8.30; python_version >= '3.6'
distlib==0.3.9
filelock==3.16.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pbr==6.1.0; python_version >= '2.6'
pipenv==2024.4.0; python_version >= '3.8'
pipenv-to-requirements==0.9.0
platformdirs==4.3.6; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-telegram-bot==21.7; python_version >= '3.9'
setuptools==75.5.0; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
virtualenv==20.27.1; python_version >= '3.8'
yt-dlp==2024.11.4; python_version >= '3.9'
